Don't terminate on SIGPIPE.
This is to accommodate applications that call TPC and substitute its output
with a pipe. TPC needs to be able to shut down in an ordered fashion when
the caller exits and TPC tries to write to a broken pipe.

Appropriate mechanism is already in place (see r225), except SIGPIPE stands
in its way on Linux.

This change removes SIGPIPE from the picture.